Move search results logic
